# Optimise a loop in Bitmap::SetBits.