@Smattr 2016-06-18
Just wanted to give this one a gentle bump in case it's slipped through the cracks. If it's already on your radar, apologies for the noise. Also, if you'd prefer this to go through some discussion on the mailing list I can post it there as well. Thanks.