@dcjm 2016-06-18
I hadn't forgotten this but there were some issues that needed a bit of consideration.  Bitmaps are used in several situations but the most important is during the garbage-collection marking phase.  In typical ML programs the vast majority of cells are small, pairs or triples, so it's important to be able to set small numbers of bits with the minimum cost.  Larger cells are much less frequent so the advantage of being able to set multiple bytes isn't that significant. Indeed the cost of the function call to memset could outweigh the saving if only a single byte is being set.
On the question of `vector<bool>`: it might be possible but I'm slightly concerned about what would happen if there was insufficient memory for a bitmap.  That's actually a frequent occurrence, particularly with Isabelle in 32-bit mode.  It could be worth experimenting.
David