@Smattr 2016-06-19
Thanks. Sorry for the unnecessary prompting ;)

Good point about the small cells. I knew this was the case, but hadn't thought through the consequences of this. We could special case short writes? My local libc (x86_64 Glibc) special cases some calls to `memset`, but it looks like this only ever triggers when the third argument is constant. My thinking when posting this PR was that a platform-specific optimised `memset` is likely to beat any more general inline implementation. However, I was not especially thinking about the case of small lengths.

With respect to the `vector<bool>` question, I was not thinking of allocation on demand, but still allocating all memory upfront. Something along the lines of calling `std::vector::reserve` in place of `Bitmap::Create`. It would seem callers of `Bitmap::Create` already need to anticipate the possibility of `calloc` failing. I agree that allocating on demand and thereby forcing callers to anticipate failure any time they set parts of the bitmap would be pretty unpleasant. Again, my thinking here was that replacing this with `vector<bool>` would reduce the amount of code to maintain in PolyML and allow taking advantage of platform-specific optimisations implemented in the C++ standard library. I'll make these changes locally and take some measurements of something like l4.verified to try and quantify a bit more accurately what (if any) benefit there is to be had here.