@Smattr 2016-06-23
I replaced the guts of `Bitmap` with an implementation using `vector<bool>`. I just did the obvious thing for all functions and didn't change any function visibility (no moving bitmap.cpp functions into bitmap.h or vice versa). I didn't attempt to make the changes neat, but they should be representative of the final result, performance-wise. I can send you the exact diff if you're curious.

Anyway, the punchline is that this made no appreciable difference. To evaluate this, I backported the changes to PolyML 5.6-1 that's used by my local Isabelle/HOL and built a fairly intensive l4.verified session, CRefine. Duration was around 2 hours both before and after the changes, with the difference not being significant.

Not the result I expected, but still nice to have quantifiable evidence that the current implementation is pretty close to optimal :)