@dcjm 2016-06-23
Thank you for doing that.  It's interesting to have those results.  Looked at another way, it means that changing to using vector<bool> would not make anything worse.  So if there are advantages in maintainability it could be worthwhile.  Bitmaps are less significant than they used to be.  At one stage, prior to version 5.5 I think, they were used in minor collections as well as major, now they are just used in the, less frequent, major collections.
It's certainly worth thinking about this, though I'd want to check that failure to allocate was handled properly.
I have wondered about using std::vector for some of the tables in the MemMgr.  It might be cleaner than the existing code.