@Smattr 2016-06-27
Good point. I've pushed the changes I made to https://github.com/Smattr/polyml/commit/54a3f9d6960d146f8c787d0bd8fa1d81dabaad40. To merge them, I imagine the Autotools stuff might need to be tweaked to also check stdlib support. What do you think?